pyproject.toml
[project]
name = "my-discord-bot"
version = "4.2.0"
description = "Discord Bot by Reguna"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.*",
    "aiosqlite==0.*",
    "alembic==1.*",
    "discord-py[voice]==2.*",
    "gallery-dl==1.*",
    "google-api-python-client==2.*",
    "openai==1.*",
    "pygit2==1.*",
    "python-dotenv==1.*",
    "sqlalchemy[asyncio]==2.0.*",
    "wavelink==3.*",
]

[project.scripts]
my-discord-bot = "my_discord_bot:main"
get-bot-version = "my_discord_bot:get_bot_version"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["google-api-python-client-stubs", "ipykernel", "ruff"]

[tool.pyright]
typeCheckingMode = "strict"
reportCallInDefaultInitializer = true
